+++ src/main/java/de/beachboys/aoc2017/Day23.java
@@ -45,7 +45,7 @@ public class Day23 extends Day {
 
         public void runProgram(List<String> program) {
             while (instructionPointer < program.size()) {
-                instructionPointer += executeInstruction(program.get(instructionPointer));
+                instructionPointer += (int) executeInstruction(program.get(instructionPointer));
             }
         }
